Train script & DiceFocalLoss improved
- The training script has been updated to use 2D Dataset or optimized 3D Dataset, without loading all data into memory.
- Random seed is now set for training.
- Ensured that DiceFocalLoss input is Long.

+++ digipanca/src/losses/dice.py
@@ -178,8 +178,14 @@ class DiceFocalLoss(nn.Module):
 
         # Convert y_true to one-hot encoding if necessary
         if y_true.dim() == 3:   # 2D: (B, H, W) → (B, C, H, W)
-            y_true = F.one_hot(y_true, num_classes=num_classes).permute(0, 3, 1, 2).float()
+            y_true = F.one_hot(
+                y_true.long(),
+                num_classes=num_classes
+            ).permute(0, 3, 1, 2).float()
         elif y_true.dim() == 4: # 3D: (B, D, H, W) → (B, C, D, H, W)
-            y_true = F.one_hot(y_true, num_classes=num_classes).permute(0, 4, 1, 2, 3).float()
+            y_true = F.one_hot(
+                y_true.long(),
+                num_classes=num_classes
+            ).permute(0, 4, 1, 2, 3).float()
 
         return self.monai_dice_focal(y_pred, y_true)
